Change refundPercent type from double to float

+++ src/GuildHouseConfig.cpp
@@ -14,7 +14,7 @@ void GuildHouseConfig::Load()
     _enabled = sConfigMgr->GetOption<bool>("GuildHouse.Enable", false);
     _free = sConfigMgr->GetOption<bool>("GuildHouse.Free", false);
 
-    _refundPercent = sConfigMgr->GetOption<double>("GuildHouse.Refund", .8);
+    _refundPercent = sConfigMgr->GetOption<float>("GuildHouse.Refund", .8);
 
     if (_refundPercent > 1)
         _refundPercent = 1;
